Preparing Lab 1 notebook
The code in the notebook and example data was taken from Oisin's code
for the course ML for Ecologists. Translated the R code to Python to run
on Google Colab.

Repo with original code: https://github.com/macaodha/ML_Ecology

Still need to include the material in the associated slides. Perhaps
split the laboratory on different sections to mirror the slide
structure.